wlan: If DFS scan mode is set to 1 allow passive/active conversion
If DFS scan mode is set to DFS_CHNL_SCAN_ENABLED_NORMAL, active to
passive and vice versa conversions are not allowed for DFS channels.

With mode set to DFS_CHNL_SCAN_ENABLED_NORMAL, it should behave
as default mode and the fix is to allow active to passive and vice
versa conversions for DFS channels.

Change-Id: I2e5b61b768c003ea7f3fb83762aa295e333a9347
CRs-Fixed: 1011499